Move handleEffects into hooks.js so most hooks code is in hooks file
